driver entries

+++ Logic/Responder.py
@@ -9,6 +9,7 @@ import Entity.Event as Event
 import Entity.RepeatingEvent as RepeatingEvent
 import Logic.UserController as UserController
 import Logic.EventController as EventController
+import Logic.DriverController as DriverController
 import main
 import logging
 from google.appengine.ext import ndb
@@ -55,7 +56,7 @@ def parseEvent(user, additional):
 
 def respondTo(message, sender):
 
-	MAINTENANCE = True
+	MAINTENANCE = False
 
 	message_id = message.get('message_id')
 	date = message.get('date')
@@ -76,7 +77,7 @@ def respondTo(message, sender):
 		else:
 			answer = answer + u'Achtung, Wartung aktiv!\n'
 
-	user = UserController.createOrUpdate(senderID, senderFirstName, chat_id)
+	user = UserController.getOrCreate(senderID)
 	textSplit = text.partition(' ')
 	command = textSplit[0].lower()
 	additional = None
@@ -92,18 +93,25 @@ def respondTo(message, sender):
 	if command.startswith(u'info'):
 		return  answer + UserController.infoForEvent(user, additional)
 
+	if command.startswith(u'fahrtlöschen'):
+		return  answer + DriverController.delete(user, additional)
+	if command.startswith(u'fahrt'):
+		return  answer + DriverController.create(user, additional)
+
 	if user.admin:
 		if command.startswith(u'erstell'):
 			return  answer + EventController.create(user, additional)
 		if command.startswith(u'lösch'):
 			return  answer + EventController.delete(user, additional)
+		if command.startswith(u'fahrerliste'):
+			return  answer + DriverController.listByUser(additional)
 	
 
 	
 	return answer + getHelpText(user)
 
 def getHelpText(user):
-	answer=  u'Folgende befehle sind möglich: \n\t an -> anmelden \n\t ab -> abmelden \n\t info (x)-> Informationen zu einem (oder x) Event(s) \n\n Zusätzlich kann ein bestimmter Tag (z.B. "Montag") oder ein Datum (TT.MM.JJJJ) angegeben werden. Wird nichts angegeben, wird das nächste Training/Spiel genommen.'
+	answer=  u'Folgende befehle sind möglich: \n\t an -> anmelden \n\t ab -> abmelden \n\t info (x)-> Informationen zu einem (oder x) Event(s) \n\n Zusätzlich kann ein bestimmter Tag (z.B. "Montag") oder ein Datum (TT.MM.JJJJ) angegeben werden. Wird nichts angegeben, wird das nächste Training/Spiel genommen.\n\n Um Fahrten einzutragen fahrt TT.MM.JJJJ mit dem Datum des jeweiligen spiels eingeben. Einträge können mit fahrtlöschen TT.MM.JJJJ wieder gelöscht werden.'
 	if user.admin:
-		answer = answer + u'\n\n Admin befehle: \n\n erstelle Name;(Ort);Datum/Tag;Zeit -> Ein Event erstellen. Angabe von einem Ort ist optional. Wird ein Datum (TT.MM.JJJJ) angegeben, wird ein einzelnes Event erstellt. Wird ein Wochentag angegeben, muss zusätzlich ein Enddatum angegeben werden. Bis zu diesem wird das Event dann jede Woche wiederholt. Die Zeit muss im Format HH:MM angegeben werden. ";" als Trennzeichen, damit auch Leerzeichen möglich sind. Pro Tag ist nur ein Event Möglich. \n \n lösche -> Ein einzelnes Event löschen. Welches wird so angegeben wie bei anmelden/abmelden/info (Wochentag, Datum oder nichts)'
+		answer = answer + u'\n\n Admin befehle: \n\n erstelle Name;(Ort);Datum/Tag;Zeit -> Ein Event erstellen. Angabe von einem Ort ist optional. Wird ein Datum (TT.MM.JJJJ) angegeben, wird ein einzelnes Event erstellt. Wird ein Wochentag angegeben, muss zusätzlich ein Enddatum angegeben werden. Bis zu diesem wird das Event dann jede Woche wiederholt. Die Zeit muss im Format HH:MM angegeben werden. ";" als Trennzeichen, damit auch Leerzeichen möglich sind. Pro Tag ist nur ein Event Möglich. \n \n lösche -> Ein einzelnes Event löschen. Welches wird so angegeben wie bei anmelden/abmelden/info (Wochentag, Datum oder nichts).  \n\n fahrerliste TT.MM.JJJJ-> Fahrten ab angegebenem Datum auflisten.'
 	return answer
